prj_setup.sh: Move project specific package install to prj_setuo.sh
The project requires git and pip etc so put these in prj_setup.sh.
prj_setup should work in other contexts that just the this Dockerfile.

Use sudo for package install for setup_common.
The old Dockerfile always ran as root so w/ & w/o sudo were really the
same.  That is not true in the general case.

Signed-off-by: Bill Mills <bill.mills@linaro.org>