prj_setup.sh: Disable ZEPHYR_SDK_INSTALL_DIR
This is not being used.

As things stand now the SDK gets installed in a directory of the
project top dir.

The variable said it was going to /opt but sudo is not being used to
install nor remove this dir.

Disable these lines for now.  Remove them once we detect an already
installed SDK.

Signed-off-by: Bill Mills <bill.mills@linaro.org>