HACK: prj_setup.sh: add ensure_local_bin
This is WIP but checkin this version.
Make a function to ensure we get ~/.local/bin in the PATH both for this
shell and for future shells.

This is messy due to ubuntu not putting it in the path before it exists
and skipping .bashrc for non-interactive paths.

I reality none of these .bashrc changes are needed for ubuntu, it will work
next login.  You only need to get it in the path for this time.

Signed-off-by: Bill Mills <bill.mills@linaro.org>